Removing root dir path from unit tests configuration
