+++ init.py
@@ -6,25 +6,9 @@ All rights reserved
 """
 import sys
 from neurodamus import commands
+from neurodamus.utils.cli import extract_arguments
 from neuron import h
 import logging
-
-
-def extract_arguments(args):
-    """Get the options for neurodamus and launch it.
-
-    We can't use positional arguments with special so we look for
-    --configFile=FILE, which defaults to simulation_config.json
-    """
-    config_file = "simulation_config.json"
-    for i, arg in enumerate(args[1:]):
-        if not arg.startswith("-"):
-            raise ValueError("Positional arguments are not supported")
-        elif arg.startswith("--configFile="):
-            config_file = arg.split('=')[1]
-            break
-    args = [config_file] + [x for x in args[1:] if not x.startswith("--configFile=")]
-    return args
 
 
 def main():
@@ -34,7 +18,7 @@ def main():
     except ValueError:
         logging.error("Positional arguments are not supported by init.py; please specify --configFile=<config> "
                       "to run this script (or leave empty to use the default, ./simulation_config.json).")
-        h.quit(1)
+        return 1
 
     return commands.neurodamus(args)
